+++ src/controllers/campus/campusPermissionChecks.js
@@ -1,0 +1,43 @@
+// O========================================================================================O
+
+/*
+    O==============================================================================O
+    |   Funções de controle para verificar as permissões dos usuários em campus    |
+    O==============================================================================O
+
+    Lista de funções:  
+    - [x] checkUsertoEdit;
+*/
+
+// O========================================================================================O
+
+// Importando módulos:
+
+// Módulo dos Models Read de usuário:
+import UserRead from "../../models/user/userOperations/userReadModels";
+
+// O========================================================================================O
+
+// Função para verificar se o usuário tem permissão para editar um campus:
+const checkUserToEdit = async (userId) => {
+  // Recuperando o nível de administração do usuário no campus:
+  const user = await UserRead.getUserByID(userId);
+
+  // Verificando se o usuário existe e se ele tem permissão para editar um campus:
+  if (user.status === false || user.userData.campusAdminLevel < 3) {
+    return { status: false };
+  }
+
+  return {
+    status: true,
+    userData: user.userData,
+  };
+};
+
+// O========================================================================================O
+
+export default {
+  checkUserToEdit,
+};
+
+// O========================================================================================O
